+++ local_modules/google-sheets.js
@@ -1,0 +1,181 @@
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Other Google sheets...
+// copied from https://developers.google.com/sheets/quickstart/nodejs
+// we'll need to refactor and clean up
+
+var fs = require('fs');
+var readline = require('readline');
+var google = require('googleapis');
+var googleAuth = require('google-auth-library');
+
+
+var GoogleSheet = function () {
+
+ 
+  // If modifying these scopes, delete your previously saved credentials
+  // at ~/.credentials/sheets.googleapis.com-nodejs-quickstart.json
+  var SCOPES = ['https://www.googleapis.com/auth/spreadsheets'];
+  var TOKEN_DIR = 'credentials/';
+  var TOKEN_PATH = 'credentials/sheets_token.json';
+
+  // Load client secrets from a local file.
+  fs.readFile('credentials/client_secret.json', function processClientSecrets(err, content) {
+    if (err) {
+      console.log('Missing credentials/client_secret.json please visit https://developers.google.com/sheets/quickstart/nodejs and read "Turn on the Google Sheets API"' + err);
+      return;
+    }
+    // Authorize a client with the loaded credentials, then call the
+    // Google Sheets API.
+    authorize(JSON.parse(content), listMajors);
+  });
+
+  /**
+   * Create an OAuth2 client with the given credentials, and then execute the
+   * given callback function.
+   *
+   * @param {Object} credentials The authorization client credentials.
+   * @param {function} callback The callback to call with the authorized client.
+   */
+  function authorize(credentials, callback) {
+    var clientSecret = credentials.web.client_secret;
+    var clientId = credentials.web.client_id;
+    var redirectUrl = credentials.web.redirect_uris[0];
+    var auth = new googleAuth();
+    var oauth2Client = new auth.OAuth2(clientId, clientSecret, redirectUrl);
+
+    // Check if we have previously stored a token.
+    fs.readFile(TOKEN_PATH, function(err, token) {
+      if (err) {
+        getNewToken(oauth2Client, callback);
+      } else {
+        oauth2Client.credentials = JSON.parse(token);
+        callback(oauth2Client);
+      }
+    });
+  }
+
+  /**
+   * Get and store new token after prompting for user authorization, and then
+   * execute the given callback with the authorized OAuth2 client.
+   *
+   * @param {google.auth.OAuth2} oauth2Client The OAuth2 client to get token for.
+   * @param {getEventsCallback} callback The callback to call with the authorized
+   *     client.
+   */
+  function getNewToken(oauth2Client, callback) {
+    var authUrl = oauth2Client.generateAuthUrl({
+      access_type: 'offline',
+      scope: SCOPES
+    });
+    console.log('Authorize this app by visiting this url: ', authUrl);
+    var rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout
+    });
+    rl.question('Enter the code from that page here: ', function(code) {
+      rl.close();
+      oauth2Client.getToken(code, function(err, token) {
+        if (err) {
+          console.log('Error while trying to retrieve access token', err);
+          return;
+        }
+        oauth2Client.credentials = token;
+        storeToken(token);
+        callback(oauth2Client);
+      });
+    });
+  }
+
+  /**
+   * Store token to disk be used in later program executions.
+   *
+   * @param {Object} token The token to store to disk.
+   */
+  function storeToken(token) {
+    try {
+      fs.mkdirSync(TOKEN_DIR);
+    } catch (err) {
+      if (err.code != 'EEXIST') {
+        throw err;
+      }
+    }
+    fs.writeFile(TOKEN_PATH, JSON.stringify(token));
+    console.log('Token stored to ' + TOKEN_PATH);
+  }
+
+  /**
+   * Print the names and majors of students in a sample spreadsheet:
+   * https://docs.google.com/spreadsheets/d/1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgvE2upms/edit
+   */
+  function listMajors(auth) {
+    var sheets = google.sheets('v4');
+    sheets.spreadsheets.values.get({
+      auth: auth,
+      spreadsheetId: '1Xm-mmMthNuE0H2hjc2x5tbfAJSDBgp7Gv0ErOeBZiYc',
+      range: 'Sheet1!A1:D4',
+    }, function(err, response) {
+      if (err) {
+        console.log('The API returned an error: ' + err);
+        return;
+      }
+      var rows = response.values;
+      if (rows == undefined || rows.length == 0) {
+        console.log('No data found.');
+      } else {
+        console.log('rows:');
+        for (var i = 0; i < rows.length; i++) {
+          var row = rows[i];
+          // Print columns A and E, which correspond to indices 0 and 4.
+          console.log('%s, %s', row[0], row[3]);
+        }
+      }
+    });
+  } 
+
+  //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+  // Adding to the above example with what we need 
+
+
+  // adding append funtion
+  function auth_append(auth, arr) {
+    var sheets = google.sheets('v4');
+    sheets.spreadsheets.values.append({
+      auth: auth,
+      spreadsheetId: config.GOOGLE_DOC_KEY,
+      range: 'Sheet1',
+      valueInputOption: 'USER_ENTERED',
+      resource: { 
+          "range": "Sheet1",
+          majorDimension: "ROWS",
+          "values":
+             [arr]
+      },
+    },
+     function(err, response) {
+      if (err) {
+        console.log('The API returned an error: ' + err);
+        return;
+      }
+      console.log('append response: ' + JSON.stringify(response));
+    });
+  } 
+
+  GoogleSheet.prototype.append = function append(arr) { 
+    fs.readFile('credentials/client_secret.json', function processClientSecrets(err, content) {
+      if (err) {
+        console.log('Error loading client secret file: ' + err);
+        return;
+      }
+      // Authorize a client with the loaded credentials, then call the
+      // Google Sheets API.
+      authorize(JSON.parse(content), function(auth) {auth_append(auth,arr)});
+    });
+  }
+
+};
+
+exports.GoogleSheet = GoogleSheet;
+ 
+
+
+
